LayoutTests/svg/custom/svg-parse-overflow-4-expected.txt

CONSOLE MESSAGE: line 9: Error: Invalid value for <foreignObject> attribute y="105.0e50"
